stock_elt_project/requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioitertools==0.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
botocore==1.38.27
duckdb==1.3.1
frozenlist==1.7.0
fsspec==2025.5.1
idna==3.10
jmespath==1.0.1
multidict==6.5.0
numpy==2.2.6
pandas==2.3.0
propcache==0.3.2
py4j==0.10.9.9
pyarrow==20.0.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
s3fs==2025.5.1
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.1
boto3
python-dotenv
psycopg2
